Alerts::class moved to Cookies for performance

+++ zFramework/Core/Facades/JustOneTime.php
@@ -4,8 +4,8 @@ namespace zFramework\Core\Facades;
 
 class JustOneTime
 {
-
     static private $session_name = 'just-one-time';
+    static $used = false;
 
     /**
      * Set just one time data.
@@ -15,6 +15,7 @@ class JustOneTime
      */
     public static function set(string $name, mixed $value): self
     {
+        self::$used = true;
         return Session::callback(function () use ($name, $value) {
             $_SESSION[self::$session_name][$name] = $value;
             return new self();
@@ -28,17 +29,15 @@ class JustOneTime
      */
     public static function get(string $name): mixed
     {
-        return Session::callback(function () use ($name) {
-            return @$_SESSION[self::$session_name][$name];
-        });
+        return Session::callback(fn() => @$_SESSION[self::$session_name][$name]);
     }
 
     /**
      * Unset All Data.
      * @return void
      */
-    public static function unset()
+    public static function unset(): void
     {
-        Session::delete(self::$session_name);
+        if (self::$used) Session::delete(self::$session_name);
     }
 }
